+++ src/test/java/school/redrover/BreadCrumbsDropDownMenuTest.java
@@ -3,7 +3,7 @@ package school.redrover;
 import org.testng.Assert;
 import org.testng.annotations.Test;
 import school.redrover.common.BaseTest;
-import school.redrover.page.FreestyleProjectPage;
+import school.redrover.page.freestyle.FreestyleProjectPage;
 import school.redrover.page.HomePage;
 
 public class BreadCrumbsDropDownMenuTest extends BaseTest {
