+++ src/test/java/school/redrover/Folder2Test.java
@@ -6,7 +6,7 @@ import org.testng.annotations.Ignore;
 import org.testng.annotations.Test;
 import school.redrover.common.BaseTest;
 import school.redrover.page.HomePage;
-import school.redrover.page.OrganizationFolderPage;
+import school.redrover.page.organizaitionfolder.OrganizationFolderPage;
 
 public class Folder2Test extends BaseTest {
