+++ src/test/java/school/redrover/FreestyleProjectConfigurationBuildTriggersTest.java
@@ -9,7 +9,7 @@ import org.testng.Assert;
 import org.testng.annotations.Test;
 import school.redrover.common.BaseTest;
 import school.redrover.common.TestUtils;
-import school.redrover.page.FreestyleConfigurationPage;
+import school.redrover.page.freestyle.FreestyleConfigurationPage;
 import school.redrover.page.HomePage;
 
 import java.util.List;
