+++ src/test/java/school/redrover/FreestyleProjectTest.java
@@ -1,12 +1,9 @@
 package school.redrover;
 
-import org.openqa.selenium.By;
-import org.openqa.selenium.support.ui.ExpectedConditions;
 import org.testng.Assert;
 import org.testng.annotations.Test;
 import school.redrover.common.BaseTest;
-import school.redrover.page.FreestyleConfigurationPage;
-import school.redrover.page.FreestyleProjectPage;
+import school.redrover.page.freestyle.FreestyleProjectPage;
 import school.redrover.page.HomePage;
 
 import java.util.List;
