+++ src/test/java/school/redrover/NewItem5Test.java
@@ -8,9 +8,8 @@ import org.testng.annotations.Ignore;
 import org.testng.annotations.Test;
 import school.redrover.common.BaseTest;
 import school.redrover.page.HomePage;
-import school.redrover.page.NewItemPage;
+import school.redrover.page.newitam.NewItemPage;
 
-import java.util.ArrayList;
 import java.util.List;
 
 public class NewItem5Test extends BaseTest {
