+++ src/test/java/school/redrover/OrganizationFolderTest.java
@@ -9,7 +9,7 @@ import org.testng.annotations.Ignore;
 import org.testng.annotations.Test;
 import school.redrover.common.BaseTest;
 import school.redrover.page.HomePage;
-import school.redrover.page.OrganizationFolderPage;
+import school.redrover.page.organizaitionfolder.OrganizationFolderPage;
 
 public class OrganizationFolderTest extends BaseTest {
