+++ src/test/java/school/redrover/Pipeline1Test.java
@@ -9,7 +9,7 @@ import org.testng.Assert;
 import org.testng.annotations.Test;
 import school.redrover.common.BaseTest;
 import school.redrover.page.HomePage;
-import school.redrover.page.PipelineProjectPage;
+import school.redrover.page.pipeline.PipelineProjectPage;
 
 import java.time.Duration;
