+++ src/test/java/school/redrover/PipelineTest.java
@@ -4,8 +4,8 @@ import org.testng.Assert;
 import org.testng.annotations.Test;
 import school.redrover.common.BaseTest;
 import school.redrover.page.HomePage;
-import school.redrover.page.PipelineConfigurationPage;
-import school.redrover.page.PipelineProjectPage;
+import school.redrover.page.pipeline.PipelineConfigurationPage;
+import school.redrover.page.pipeline.PipelineProjectPage;
 
 public class PipelineTest extends BaseTest {
