+++ src/test/java/school/redrover/User1Test.java
@@ -5,7 +5,7 @@ import org.testng.Assert;
 import org.testng.annotations.Test;
 import school.redrover.common.BaseTest;
 import school.redrover.page.HomePage;
-import school.redrover.page.UsersPage;
+import school.redrover.page.user.UsersPage;
 
 import java.util.List;
