+++ src/test/java/school/redrover/page/account/AccountPage.java
@@ -1,4 +1,4 @@
-package school.redrover.page;
+package school.redrover.page.account;
 
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
