+++ src/test/java/school/redrover/page/buildhistory/BuildHistoryPage.java
@@ -1,4 +1,4 @@
-package school.redrover.page;
+package school.redrover.page.buildhistory;
 
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
