+++ src/test/java/school/redrover/page/folder/FolderProjectPage.java
@@ -1,4 +1,4 @@
-package school.redrover.page;
+package school.redrover.page.folder;
 
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
