+++ src/test/java/school/redrover/page/multiconfiguration/MultibranchConfigurationPage.java
@@ -1,4 +1,4 @@
-package school.redrover.page;
+package school.redrover.page.multiconfiguration;
 
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
