+++ src/test/java/school/redrover/page/newitam/NewItemWithinFolderPage.java
@@ -1,4 +1,4 @@
-package school.redrover.page;
+package school.redrover.page.newitam;
 
 import org.openqa.selenium.By;
 import org.openqa.selenium.TimeoutException;
@@ -7,6 +7,7 @@ import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.ui.ExpectedConditions;
 import org.openqa.selenium.support.ui.WebDriverWait;
 import school.redrover.common.BasePage;
+import school.redrover.page.HomePage;
 
 import java.time.Duration;
