+++ src/test/java/school/redrover/page/organizaitionfolder/OrganizationFolderConfigurePage.java
@@ -1,4 +1,4 @@
-package school.redrover.page;
+package school.redrover.page.organizaitionfolder;
 
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
