+++ src/test/java/school/redrover/page/user/UserAdminPage.java
@@ -1,4 +1,4 @@
-package school.redrover.page;
+package school.redrover.page.user;
 
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
